Input Module UI and Server

This module provides the input interface and server logic for selecting city, weather parameters, and running simulations.

inputModuleUI(id)

Arguments

id

A unique identifier for the module.

Value

Reactive values including `is_valid_city`, `selected_city`, `parameter`, `simulations`, and `run_simulation`.

Examples

# Example usage:
# In UI:
inputModuleUI("location1")
#> <div class="form-group shiny-input-container">
#>   <label class="control-label" id="location1-city-label" for="location1-city">City</label>
#>   <input id="location1-city" type="text" class="shiny-input-text form-control" value="Lausanne"/>
#> </div>
#> <div class="form-group shiny-input-container">
#>   <label class="control-label" id="location1-parameter-label" for="location1-parameter">Parameter</label>
#>   <div>
#>     <select id="location1-parameter" class="shiny-input-select"><option value="temperature" selected>Temperature</option>
#> <option value="humidity">Humidity</option>
#> <option value="pressure">Pressure</option></select>
#>     <script type="application/json" data-for="location1-parameter" data-nonempty="">{"plugins":["selectize-plugin-a11y"]}</script>
#>   </div>
#> </div>
#> <button id="location1-submit" type="button" class="btn btn-default action-button">Submit</button>
#> <div class="form-group shiny-input-container">
#>   <label class="control-label" id="location1-simulations-label" for="location1-simulations">Number of Simulations</label>
#>   <input id="location1-simulations" type="number" class="shiny-input-number form-control" value="10000" min="1"/>
#> </div>
#> <button id="location1-run_simulation" type="button" class="btn btn-default action-button">
#>   <i class="fas fa-play" role="presentation" aria-label="play icon"></i>
#>   Run Simulation
#> </button>
# In Server:
inputModule(input, output, session, API_KEY)
#> $is_valid_city
#> reactiveVal: [1] "FALSE" 
#> 
#> $selected_city
#> reactiveVal: [1] "NULL" 
#> 
#> $parameter
#> reactive({
#>     input$parameter
#> }) 
#> 
#> $simulations
#> reactive({
#>     input$simulations
#> }) 
#> 
#> $run_simulation
#> reactive({
#>     input$run_simulation
#> }) 
#>